Move BlobRef
It's not needed in image layers anymore, so move it into delta_layer.rs
