+++ pageserver/src/layered_repository/block_io.rs
@@ -1,8 +1,6 @@
 //!
 //! Low-level Block-oriented I/O functions
 //!
-//!
-//!
 
 use crate::page_cache;
 use crate::page_cache::{ReadBufResult, PAGE_SZ};
